+++ Project Soul/Assets/Scripts/Stratos/AI/GoapPlanner.cs
@@ -2,13 +2,13 @@
 using System.Collections.Generic;
 using UnityEngine;
 
-/**
+/*
  * Plans what actions can be completed in order to fulfill a goal state.
  */
 public class GoapPlanner
 {
 
-    /**
+    /*
 	 * Plan what sequence of actions can fulfill the goal.
 	 * Returns null if a plan could not be found, or a list of the actions
 	 * that must be performed, in order, to fulfill the goal.
@@ -84,7 +84,7 @@ public class GoapPlanner
         return queue;
     }
 
-    /**
+    /*
 	 * Returns true if at least one solution was found.
 	 * The possible paths are stored in the leaves list. Each leaf has a
 	 * 'runningCost' value where the lowest cost will be the best action
@@ -128,7 +128,7 @@ public class GoapPlanner
         return foundOne;
     }
 
-    /**
+    /*
 	 * Create a subset of the actions excluding the removeMe one. Creates a new set.
 	 */
     private HashSet<GoapAction> actionSubset(HashSet<GoapAction> actions, GoapAction removeMe)
@@ -142,7 +142,7 @@ public class GoapPlanner
         return subset;
     }
 
-    /**
+    /*
 	 * Check that all items in 'test' are in 'state'. If just one does not match or is not there
 	 * then this returns false.
 	 */
@@ -166,7 +166,7 @@ public class GoapPlanner
         return allMatch;
     }
 
-    /**
+    /*
 	 * Apply the stateChange to the currentState
 	 */
     private HashSet<KeyValuePair<string, object>> populateState(HashSet<KeyValuePair<string, object>> currentState, HashSet<KeyValuePair<string, object>> stateChange)
@@ -207,7 +207,7 @@ public class GoapPlanner
         return state;
     }
 
-    /**
+    /*
 	 * Used for building up the graph and holding the running costs of actions.
 	 */
     private class Node