+++ Project Soul/Assets/Scripts/Stratos/AI_Data/Ranger.cs
@@ -71,7 +71,7 @@ public abstract class Ranger : MonoBehaviour, IGoap
     /**
 	 * Implement in subclasses
 	 */
-    public abstract HashSet<KeyValuePair<string, object>> createGoalState();
+    public abstract HashSet<KeyValuePair<string, object>> createGoalState(int goalGeneratorID);
 
 
     public void planFailed(HashSet<KeyValuePair<string, object>> failedGoal)